Extract function from Util
