+++ vitaes-webapp/src/utils/copyElement.js
@@ -1,0 +1,3 @@
+export default function copyElement(element) {
+  return JSON.parse(JSON.stringify(element));
+}
