+++ vitaes-webapp/src/utils/getHostname.js
@@ -1,0 +1,7 @@
+export default function getHostname() {
+  let hostname = `${window.location.hostname}:5000`;
+  if (hostname === 'vitaes.io:5000') {
+    hostname = 'renderer.vitaes.io';
+  }
+  return hostname;
+}
