+++ vitaes-webapp/src/utils/removeDisabled.js
@@ -1,0 +1,11 @@
+import copyElement from './copyElement'
+
+export default function removeDisabled(rawCv) {
+  const cv = copyElement(rawCv);
+  Object.keys(cv).forEach((key) => {
+    if (Array.isArray(cv[key])) {
+      cv[key] = cv[key].filter(element => !element.disable);
+    }
+  });
+  return cv;
+}
