+++ vitaes-webapp/src/utils/validateEmail.js
@@ -1,0 +1,4 @@
+export default function validateEmail(email) {
+  const re = /\S+@\S+\.\S+/;
+  return re.test(email);
+}
